Feature(#9):Moved ErrorDisplay as common component
